+++ NEWS.md
@@ -4,6 +4,6 @@
 
 # NetworkExtinction 0.1.2
 
-* Fixed parameters in `degree_distribution`, the truncated parameters were not present
 * Fixed parameters in `degree_distribution`, the intercept for power law was not present
 * Fixed `degree_distribution` to select one model per family
+* Eliminated *Truncated* distribution from `degree_distribution` since it didn't have any theoretical support